+++ packages/video_player/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.4.9
+
+* Fix event channel issue, sending messages from native to Flutter on the platform thread.
+
+## 2.4.8
+
+* Disable screensaver when player is playing.
+
 ## 2.4.7
 
 * Update pigeon to 10.0.0.